Added skeleton for new character
The character is called Warforged, but otherwise the setup is not specialized for the new character. This setup really just made it possible to select the character for a new game. New content such as cards must be added for playability, and images/animations must be added/changed to customize the look and feel.