Added basic Strike and Defend cards
The basic starting cards for Strike and Defend have been added and are working. They still just use the default images and such, and they don't yet appear in the starting deck, but they work.

This also includes the changes to the mod class that add/register cards to the game.