Patched custom hooks for when energy changes
I patched the EnergyPanel class with postfixes that simply call my hook, onEnergyChanged, once the character's energy changes. The postfixes were applied to the end of these three methods of the EnergyPanel class:
setEnergy
addEnergy
useEnergy

Later, I can amend this hook to create a mechanic that reacts to the energy changing. It is not truly a hook, though, because it is not meant to be overwritten like the onAttack method of AbstractPower. It simply reacts to the energy change so that I can call stuff from that method.