Adding logic to CrackedAetherheartRelic (WIP)
Most of the logic for the Cracked Aetherheart Relic is there, but it's not quite done yet.

This commit includes a new GainEnergyAction specific to the Warforged which does exactly the same as the normal action, plus tracks whether the card was triggered by a card being played via the Cracked Aetherheart and, if so, calls its aetherburn method.

Turns out that I might not need the onEnergyGained "hook" I patched in before, since following that intended design would require even more complex patches, possibly destructive ones, and the result (while logical) would be difficult for some players to follow. This way is simple for both development and gameplay strategy.