RelicStrings and moved 0-energy-per-turn to relic
The Warforged now has a 3-energy-per-turn default again, which is overwritten by the CrackedAetherheartRelic_Warforged back down to 0, since it only makes sense alongside the other mechanics on the relic.

I also edited the RelicStrings to explain (most of) the mechanics on the aetherheart relic. The energy-based Dex/Str mechanic is not shown, mostly because there is not enough  room on the tooltip for it, but also because if I can separate it onto a second relic, that would be much easier to replace with an upgraded version down the line than the monolithic relic I have now. That way, I could leave the core mechanics alone and only alter a secondary mechanic with the upgraded relic.