Progress on Biocircuitry
The logic for the Biocircuitry relic is there, and should work since it's copied from the Cracked Aetherheart relic where it was built, but the way that Balance needed to be called meant adjusting my custom GainEnergyAction_Warforged class as well. I have not yet tested it. Added item to RelicStrings