Added Living Construct relic
The third and final starter relic (although the first in the list shown in game) is done. The mechanic behind the character gaining/losing no energy between turns has been successfully moved from the Cracked Aetherheart relic to the Living Construct relic.

This did manage to spread out the language a bit for the relics' mechanics, but the Cracked Aetherheart is still a bit too long in the examine view. I don't think there is anything I can do about that without adding a scroll wheel or something to the description section of the examine view. At least it looks fine in the mouseover tooltip in-game where most people will see it.