New starter relic
I added the DiagnosticsRelic_Warforged starter relic and deleted the LivingConstruct relic. Now, the Living Construct, Aetherburn, and Volatile Aether mechanics are all back on the Cracked Aetherheart relic, and the Overload mechanic was moved to a new Uncommon relic, the QuantumTransistorRelic_Warforged.

The Diagnostics relic is not finished, but the bones are set up.