Diagnostics onEquip/onUnequip support
The diagnostics relic now handles being equipped and unequipped, raising and lowering max HP appropriately based on the number of unique replacement part cards in the master deck at the time.

This included adding a new function to GeneralUtils that will tell you if a card group (such as the player's master deck) already has a card with the same name as the card in question.

Lastly, this includes an override in TheWarforged.java (character class) of the decreaseMaxHealth method so that the amount of HP lost will appear above the character's head in red letters, similar to the green text that appears when max hp is increased. This was added because the character loses max HP if the Diagnostics relic is lost after it has granted max HP.